refactor: reorganize Payment and Product management view models into separate files
