+++ src/renderer/src/pages/PaymentPage.jsx
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types'
-import { usePaymentViewModel } from '../hooks/usePaymentViewModel'
+import { usePaymentViewModel } from '../hooks/usePayment'
 import CardIcon from '../assets/Icons/Card.png'
 import CashIcon from '../assets/Icons/Cash.png'
 import PersonIcon from '../assets/Icons/Person.png'
